core/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1